autoselect pico i2c port

+++ modules/files/preference/esp/modPreference.c
@@ -41,7 +41,7 @@
 #endif
 
 
-extern uint32_t _MODPREF_start;
+extern uint8_t _MODPREF_start;
 
 #define kPreferencesStartOffset ((uintptr_t)&_MODPREF_start - (uintptr_t)kFlashStart)
 #define kPreferencesMagic 0x81213141
